move bot's ready func into own file
